requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.5.30=py36h06a4308_0
cloudpickle=2.0.0=pyhd3eb1b0_0
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=5.1.0=pyhd3eb1b0_0
efficientnet-pytorch=0.7.1=pypi_0
expat=2.4.1=h2531618_2
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
glib=2.63.1=h5a9c865_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=he6710b0_3
imageio=2.9.0=pyhd3eb1b0_0
intel-openmp=2021.3.0=h06a4308_3350
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py36h2531618_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
libedit=3.1.20210714=h7f8727e_0
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
mat4py=0.5.0=pypi_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
munch=2.5.0=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
networkx=2.5=py_0
ninja=1.10.2=hff7bd54_1
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
opencv-python=3.4.2.16=pypi_0
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f8727e_0
pcre=8.45=h295c915_0
pillow=8.2.0=py36he98fc37_0
pip=21.2.2=py36h06a4308_0
pretrainedmodels=0.7.4=pypi_0
pydensecrf=1.0rc3=py36hf8a1672_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.8.2=py3.6_cuda11.1_cudnn8.0.5_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.4.1=py36h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
scikit-image=0.17.2=py36hdf5156a_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
setuptools=58.0.4=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2020.10.1=py36hdd07704_2
timm=0.4.12=pypi_0
tk=8.6.11=h1ccaba5_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.8.2=py36
torchvision=0.9.2=py36_cu111
tornado=6.1=py36h27cfd23_0
tqdm=4.62.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
wheel=0.37.0=pyhd3eb1b0_1
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
